build a _strtok() function, used _strstr() in place of start_with(), simplified other functions to make them more readable and effective

+++ read_buf.c
@@ -16,6 +16,6 @@ ssize_t read_buf(info_t *info, char *buf, size_t *i)
 		return (0);
 	r = read(info->readfd, buf, READ_BUF_SIZE);
 	if (r >= 0)
-		*i = r;
+		*i = (size_t)r;
 	return (r);
 }
